requirements.txt
fastapi==0.112.2
kubernetes==30.1.0
pytest==8.3.3
glueops-helpers @ https://github.com/GlueOps/python-glueops-helpers-library/archive/refs/tags/v0.4.1.zip